[refactor] 여행 상태 TripStatus 이름 변경 (#87)
