+++ backend/src/test/java/dev/tripdraw/domain/trip/TripTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.domain.trip;
 
-import static dev.tripdraw.domain.trip.Status.FINISHED;
+import static dev.tripdraw.domain.trip.TripStatus.FINISHED;
 import static dev.tripdraw.exception.trip.TripExceptionType.NOT_AUTHORIZED;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
