+++ backend/src/test/java/dev/tripdraw/presentation/controller/TripControllerTest.java
@@ -1,6 +1,6 @@
 package dev.tripdraw.presentation.controller;
 
-import static dev.tripdraw.domain.trip.Status.ONGOING;
+import static dev.tripdraw.domain.trip.TripStatus.ONGOING;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
